Added Question 4
Question 4 - How often do players who make the first check win the game?
Question 4.1 - How often is the first check also linked with the end of the game?
Question 4.2 - Is the result consistent upon stratification on time control? Player elo? Both?
Question 4.3 - How are the game length in plies and the ply number of the first check related?
Question 4.4 - How often does the player with a lower elo make the first check?
Question 4.5 - How is the number of checks in a game correlated with the number of plies in the game? [This will be moved to the new Question 7.]